server/requirements.txt
Flask
Flask-CORS
requests
Flask-SQLAlchemy
Authlib
python-dotenv